[0026] - ajuste de quantidade para inteiro

+++ Código Fonte/VERSAO-ATUAL/lib/src/Blocos.Entities/Interno.Extensoes/Caracterizacoes/ModuloInformacaoCorte/InformacaoCorteDestinacao.cs
@@ -14,6 +14,6 @@ namespace Tecnomapas.Blocos.Entities.Interno.Extensoes.Caracterizacoes.ModuloInf
 		public Int32 Produto { get; set; }
 		public String ProdutoTexto { get; set; }
 
-		public Decimal Quantidade { get; set; }
+		public Int32 Quantidade { get; set; }
 	}
 }
