+++ Código Fonte/VERSAO-ATUAL/src/EtramiteX/Interno/Areas/Caracterizacoes/ViewModels/VMInformacaoCorte/InformacaoCorteResultadosVM.cs
@@ -16,7 +16,7 @@ namespace Tecnomapas.EtramiteX.Interno.Areas.Caracterizacoes.ViewModels.VMInform
 		public string DestinacaoMaterialTexto { get; set; }
 		public int Produto { get; set; }
 		public string ProdutoTexto { get; set; }
-		public Decimal Quantidade { get; set; }
+		public int Quantidade { get; set; }
 		public int Linhas { get; set; }
 	}
 }
